# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_basilex_promenade_pkg_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Error'
      message:
        type: string
      pagination:
        $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Pagination'
      status:
        type: string
    type: object
...
  internal_contexts_warehouse_inventory_adapter_http.InventoryResponse:
    properties:
      created_at:
        type: string
      currency_code:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_restocked:
        type: string
      last_updated_by:
        type: string
      location_code:
        type: string
      location_zone:
        type: string
      notes:
        type: string
      product_id:
        type: string
      product_name:
        type: string
      quantity_available:
        type: integer
      quantity_committed:
        type: integer
      quantity_on_hand:
        type: integer
      quantity_reserved:
        type: integer
      reorder_point:
        type: integer
      reorder_quantity:
        type: integer
      sku:
        type: string
      status:
        type: string
      unit_cost_cents:
        type: integer
      updated_at:
        type: string
      version:
        type: integer
      warehouse_id:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: alexander.vasilenko@gmail.com
    name: API Support
    url: https://github.com/basilex/promenade
  description: Modern backend platform for customer management, orders, and business
    workflows with clean DDD architecture
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Promenade Platform
  version: 0.1.0
paths:
...
  /warehouse/inventory:
    get:
      description: Get paginated list of inventory items
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.InventoryResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: List inventory
      tags:
      - Inventory
    post:
      consumes:
      - application/json
      description: Create a new inventory item
      parameters:
      - description: Inventory creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.CreateInventoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.InventoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Create inventory
      tags:
      - Inventory
  /warehouse/inventory/{id}:
    delete:
      description: Soft-delete inventory item
      parameters:
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Delete inventory
      tags:
      - Inventory
    get:
      description: Get inventory item by ID
      parameters:
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.InventoryResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Get inventory by ID
      tags:
      - Inventory
    put:
      consumes:
      - application/json
      description: Update inventory item details
      parameters:
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: string
      - description: Inventory update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.UpdateInventoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.InventoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Update inventory
      tags:
      - Inventory
  /warehouse/inventory/{id}/commit:
    post:
      consumes:
      - application/json
      description: Commit stock (decreases available, increases committed)
      parameters:
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: string
      - description: Commit stock request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.CommitStockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.InventoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Commit stock
      tags:
      - Inventory
  /warehouse/inventory/{id}/receive:
    post:
      consumes:
      - application/json
      description: Receive stock into inventory (increases quantity)
      parameters:
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: string
      - description: Receive stock request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.ReceiveStockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.InventoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Receive stock
      tags:
      - Inventory
  /warehouse/inventory/{id}/release:
    post:
      consumes:
      - application/json
      description: Release reserved stock when order is cancelled (Saga compensation)
      parameters:
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: string
      - description: Release reservation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.ReleaseReservationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.InventoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Release reservation
      tags:
      - Inventory
  /warehouse/inventory/{id}/reserve:
    post:
      consumes:
      - application/json
      description: Reserve stock for an order (decreases available, increases reserved)
      parameters:
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: string
      - description: Reserve stock request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.ReserveStockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.InventoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Reserve stock
      tags:
      - Inventory
  /warehouse/inventory/location/{warehouse_id}/{location_code}:
    get:
      description: Get inventory items by warehouse ID and location code
      parameters:
      - description: Warehouse ID
        in: path
        name: warehouse_id
        required: true
        type: string
      - description: Location Code
        in: path
        name: location_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.InventoryResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Get inventory by location
      tags:
      - Inventory
  /warehouse/inventory/low-stock:
    get:
      description: Get all inventory items below reorder point
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.InventoryResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Get low stock items
      tags:
      - Inventory
  /warehouse/inventory/product/{product_id}:
    get:
      description: Get all inventory items for a specific product
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_contexts_warehouse_inventory_adapter_http.InventoryResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Get inventory by product ID
      tags:
      - Inventory
...
schemes:
- http
- https
swagger: "2.0"